+++ jsx/src/components/Props.jsx
@@ -1,0 +1,9 @@
+function Props(props) {
+  console.log(props);
+  return (
+    <div>
+      <h2>안녕! {props.name}</h2>
+    </div>
+  );
+}
+export default Props;
